added inline styling to the taskcomp | copied to the other statusses
